make tests/testwrapper.sh more verbose; find input in srcdir
Since output is redirected by the automake parallel test suite driver,
turn on both -v and -x in bash for the testwrapper.sh shell
script. This helps a ton in debugging problems with the test harness
itself.

In general, the input files are in automake's srcdir and the name of
the test includes the relative path to it (even though that's supposed
to be ./). Therefore, pass srcdir in AM_LOG_FLAGS and prepend that to
the test name as part of constructing the input file's name.